Move AppSignal-specific mock modules into their own subdirectory
