+++ test/support/appsignal/fake_diagnose_report.ex
